Adding zuul jobs for new repo
This job adds a single linters tox target.

The linters target is an aggregation of linters for
this repo including:
 - bashate

Other linters such as yamllint can be added to this
repo by later commits.

This repo is primarily GO code. The zuul and tox jobs
will be more comprehensive in later commits.

This commit also adds basic contributing and hacking docs.

Change-Id: I3b760f16043ce0f0444cd153e17aa262ab30328a
Story: 2006166
Task: 36542
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>