Upgrade docker-distribution to v2.8.2
This review will upgrade docker distribution to v2.8.2 that allow
the use of 'mintlsversion' to block the use of TLS1.0, 1.1 and
weak ciphers.

This review is also adding a server to the token auth,
that allow us to add custom configurations, as such the
tlsConfig, that is needed to block tls1.0, tls1.1 and
weak ciphers.

Test Plan:

PASS: Run build-image.
PASS: Run build-pkgs -c -p registry-token-server.
PASS: Verify if ports 9001 and 9002 are blocking tls1.0, 1.1 and
      weak ciphers.

Closes-Bug: 2043217

Change-Id: Ibd7a2d7db68ed40d180724909ec505041ea23ff7
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>